requirements.txt
click==8.1.7
colorama==0.4.6
greenlet==3.1.1
joblib==1.4.2
nltk==3.9.1
numpy==2.1.2
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
six==1.16.0
SQLAlchemy==2.0.35
textblob==0.18.0.post0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
